+++ Utilities/vtkexodus2/expp.c
@@ -264,7 +264,7 @@ int ex_put_prop (int   exoid,
 /*   store property name as attribute of property array variable */
 
      if ((ncattput (exoid, propid, ATT_PROP_NAME, NC_CHAR,
-                    strlen(prop_name)+1, (void*)prop_name)) == -1)
+                    (int)strlen(prop_name)+1, (void*)prop_name)) == -1)
      {
        exerrval = ncerr;
        sprintf(errmsg,
