+++ Utilities/vtkexodus2/expvnm.c
@@ -161,7 +161,7 @@ int ex_put_var_name (int   exoid,
    start[1] = 0;
 
    count[0] = 1;
-   count[1] = strlen(var_name) + 1;
+   count[1] = (long)strlen(var_name) + 1;
 
    if (ncvarput (exoid, varid, start, count, (void*)var_name) == -1)
    {
